backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-socketio==5.10.0
python-multipart==0.0.6
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
openai==1.14.3
httpx==0.27.2
requests==2.31.0
python-dotenv==1.0.0
websockets==12.0
